Split signIn form / Add margin into inputs for better error visualization - Split signIn Form; - Change theme for input errors (Add margin into fields to check errors)
